+++ internal/app/utils.go
@@ -1,0 +1,21 @@
+package app
+
+import (
+	"crypto/md5"
+	"crypto/rand"
+	"encoding/hex"
+)
+
+// GetMD5Hash ...
+func GetMD5Hash(text []byte) string {
+	hasher := md5.New()
+	hasher.Write(text)
+	return hex.EncodeToString(hasher.Sum(nil))
+}
+
+// RandomMD5Hash returns random md5 hash for unique conifrim links
+func RandomMD5Hash() string {
+	b := make([]byte, 16)
+	rand.Read(b)
+	return GetMD5Hash(b)
+}