+++ internal/router/auth.go
@@ -42,7 +42,9 @@ func Auth(s storage.Store) negroni.HandlerFunc {
 		uuid, _ := claims["uuid"].(string)
 
 		//check from db
-		if !s.Tokens().Any(uuid) {
+		tokenRow, tokenExist := s.Tokens().Any(uuid)
+
+		if !tokenExist {
 			userid, _ := strconv.Atoi(fmt.Sprintf("%.f", claims["user_id"]))
 			s.Tokens().Delete(userid)
 			w.WriteHeader(http.StatusUnauthorized)
@@ -52,16 +54,18 @@ func Auth(s storage.Store) negroni.HandlerFunc {
 		ctxAuthorized := claims["authorized"].(bool)
 		ctxUserID := claims["user_id"].(float64)
 		ctxSchema := fmt.Sprintf("user%v", claims["user_id"])
+		ctxTransmissionKey := tokenRow.TransmissionKey
 
 		ctx := r.Context()
 		ctxWithID := context.WithValue(ctx, "id", ctxUserID)
 		ctxWithAuthorized := context.WithValue(ctxWithID, "authorized", ctxAuthorized)
 		ctxWithSchema := context.WithValue(ctxWithAuthorized, "schema", ctxSchema)
+		ctxWithTransmissionKey := context.WithValue(ctxWithSchema, "transmissionKey", ctxTransmissionKey)
 
 		// These context variables can be accesable with
 		// ctxAuthorized := r.Context().Value("authorized").(bool)
 		// ctxID := r.Context().Value("id").(float64)
 
-		next(w, r.WithContext(ctxWithSchema))
+		next(w, r.WithContext(ctxWithTransmissionKey))
 	})
 }