+++ model/payload.go
@@ -1,0 +1,5 @@
+package model
+
+type Payload struct {
+	Data string `json:"data"`
+}